[reorganization] WIP move single substitution into separate files.
